SRT-42 최대원 게임매니저 CameraZoomInOut 매개변수 자료형 변경

+++ RedTheSettlers/Assets/2.Scripts/Managers/GameManager.cs
@@ -272,7 +272,7 @@ namespace RedTheSettlers.GameSystem
         /// 카메라 줌 인/아웃
         /// </summary>
         /// <param name="ZoomValue"></param>
-        public void CameraZoomInOut(int ZoomValue)
+        public void CameraZoomInOut(float ZoomValue)
         {
             cameraCtrl.ZoomInOut(ZoomValue);
         }
